requirements.txt
telebot==0.0.5
PyMySQL==1.1.0
httpx==0.27.0
requests==2.31.0
telethon==1.34.0
python-dotenv==1.0.1
